+++ Nevar/Trees/TreeRebalancer.cs
@@ -100,11 +100,11 @@ namespace Nevar.Trees
 
 			var fromNode = from.GetNode(from.LastSearchPosition);
 			Stream val = null;
-			int pageNum;
+			long pageNum;
 			if (fromNode->Flags.HasFlag(NodeFlags.Data))
 			{
 				val = new UnmanagedMemoryStream(from.Base + from.KeysOffsets[from.LastSearchPosition] + Constants.NodeHeaderSize + fromKey.Size, fromNode->DataSize);
-				pageNum = -1;
+				pageNum = -1L;
 			}
 			else
 			{
